\name{ioanalysis}
\alias{ioanalysis}
\title{ioanalysis}


\description{
A collection of input-output table analytical functions used to analyze \code{InputOutput} objects. See \code{\link{as.inputoutput}} for details.

For a list of available functions, see \code{help(package = "ioanalysis")}
}

\author{
John J. P. Wade
}

\seealso{
For detailed documentation and .R files see \url{https://github.com/joolman/ioanalysis}
}

